+++ Utils/ErrorStats.h
@@ -2,10 +2,10 @@
 
 struct ErrorStats
 {
-    float maxFlat   = 0;
-    float maxNagata = 0;
+    double maxFlat   = 0;
+    double maxNagata = 0;
     double avgFlat   = 0;   
     double avgNagata = 0;
-    long long samples = 0;
-    float ratio = 0;        
+    double samples = 0;
+    double ratio = 0;        
 };
